golden/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-ajax-selects==2.0.0
django-mercadopago==6.0.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.5.4
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
mercadopago==0.3.5
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
pep8==1.7.1
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
setuptools-git==1.2
setuptools-scm==4.1.2
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
